+++ lib/presentation/screens/splash_screen.dart
@@ -47,23 +47,23 @@ class _SplashScreenState extends State<SplashScreen> {
   Future<void> _loadInitialData() async {
     try {
       await PreferencesService.init();
-      
+
       final objectBoxService = await ObjectBoxService.getInstance();
       Get.put(objectBoxService, permanent: true);
-      
+
       Get.put(CategoryController(), permanent: true);
       Get.put(ThemeController(), permanent: true);
-      
+
       Get.put(ExpenseController(), permanent: true);
       Get.put(RevenueController(), permanent: true);
       Get.put(LoanController(), permanent: true);
-      
+
       Get.put(AccountController(), permanent: true);
       final accountController = Get.find<AccountController>();
       final expenseController = Get.find<ExpenseController>();
       final revenueController = Get.find<RevenueController>();
       final loanController = Get.find<LoanController>();
-      
+
       Get.put(UpdateController(), permanent: true);
 
       await Future.wait([
@@ -79,17 +79,15 @@ class _SplashScreenState extends State<SplashScreen> {
 
   Future<void> navigateToDashboard() async {
     bool isFirstLaunch = PreferencesService.isFirstLaunch();
-    
+    final updateController = Get.find<UpdateController>();
 
-    final updateController = Get.put(UpdateController(), permanent: true);
-    
     if (isFirstLaunch) {
       Get.offAll(() => const OnboardingScreen());
       return;
     }
-    
+
     final hasUpdate = await updateController.checkForUpdates();
-    
+
     if (hasUpdate) {
       Get.to(() => const UpdateScreen());
     } else {
