+++ lib/presentation/widgets/accounts/accounts_summary_card.dart
@@ -1,0 +1,118 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:intl/intl.dart';
+import 'package:mybudget/core/controllers/account_controller.dart';
+import 'package:mybudget/data/models/account_model.dart';
+
+class SummaryCards extends StatelessWidget {
+  final List<AccountModel> accounts;
+  final List<dynamic> expenses;
+  final List<dynamic> revenues;
+
+  const SummaryCards({
+    required this.accounts,
+    required this.expenses,
+    required this.revenues,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final formatter = NumberFormat.currency(locale: 'fr_FR', symbol: '€');
+    final accountController = Get.find<AccountController>();
+
+    final totalBalance = accountController.getTotalBalance();
+
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: Row(
+        children: [
+          Expanded(
+            child: _buildSummaryCard(
+              context,
+              'Solde Total',
+              totalBalance,
+              formatter,
+              totalBalance >= 0
+                  ? Colors.green.shade700
+                  : Theme.of(context).colorScheme.error,
+              totalBalance >= 0 ? Icons.arrow_upward : Icons.arrow_downward,
+            ),
+          ),
+          const SizedBox(width: 16),
+          Expanded(
+            child: _buildSummaryCard(
+              context,
+              'Transactions',
+              accountController.getTotalTransactionsCount(),
+              null,
+              Theme.of(context).colorScheme.primary,
+              Icons.compare_arrows,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildSummaryCard(
+    BuildContext context,
+    String title,
+    dynamic value,
+    NumberFormat? formatter,
+    Color color,
+    IconData icon,
+  ) {
+    final formattedValue =
+        formatter != null ? formatter.format(value) : value.toString();
+
+    return Container(
+      padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 20),
+      decoration: BoxDecoration(
+        color: Theme.of(context).colorScheme.surface,
+        borderRadius: BorderRadius.circular(16),
+        border: Border.all(
+          color: Theme.of(context).colorScheme.surfaceVariant.withOpacity(0.5),
+          width: 1,
+        ),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            title,
+            style: TextStyle(
+              fontSize: 14,
+              color: Theme.of(context).colorScheme.onSurfaceVariant,
+            ),
+          ),
+          const SizedBox(height: 8),
+          Row(
+            children: [
+              Container(
+                padding: const EdgeInsets.all(6),
+                decoration: BoxDecoration(
+                  color: color.withOpacity(0.1),
+                  shape: BoxShape.circle,
+                ),
+                child: Icon(icon, color: color, size: 18),
+              ),
+              const SizedBox(width: 12),
+              Expanded(
+                child: Text(
+                  formattedValue,
+                  style: TextStyle(
+                    fontSize: 20,
+                    fontWeight: FontWeight.bold,
+                    color: color,
+                  ),
+                  overflow: TextOverflow.ellipsis,
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
