+++ lib/presentation/widgets/common/delete_confirmation_dialog.dart
@@ -1,0 +1,85 @@
+import 'package:flutter/material.dart';
+
+class DeleteConfirmationDialog extends StatelessWidget {
+  final String title;
+  final String message;
+  final VoidCallback onConfirm;
+  final VoidCallback? onCancel;
+  final String confirmButtonText;
+  final String cancelButtonText;
+
+  const DeleteConfirmationDialog({
+    required this.title,
+    required this.message,
+    required this.onConfirm,
+    this.onCancel,
+    this.confirmButtonText = 'Supprimer',
+    this.cancelButtonText = 'Annuler',
+    super.key,
+  });
+
+  static Future<bool?> show({
+    required BuildContext context,
+    required String title,
+    required String message,
+    required VoidCallback onConfirm,
+    VoidCallback? onCancel,
+    String confirmButtonText = 'Supprimer',
+    String cancelButtonText = 'Annuler',
+  }) async {
+    return showDialog<bool>(
+      context: context,
+      builder: (context) => DeleteConfirmationDialog(
+        title: title,
+        message: message,
+        onConfirm: onConfirm,
+        onCancel: onCancel,
+        confirmButtonText: confirmButtonText,
+        cancelButtonText: cancelButtonText,
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    
+    return AlertDialog(
+      title: Text(title),
+      content: Text(message),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(16),
+      ),
+      actions: [
+        TextButton(
+          onPressed: () {
+            Navigator.of(context).pop(false);
+            if (onCancel != null) {
+              onCancel!();
+            }
+          },
+          style: TextButton.styleFrom(
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
+            ),
+          ),
+          child: Text(cancelButtonText),
+        ),
+        ElevatedButton(
+          onPressed: () {
+            Navigator.of(context).pop(true);
+            onConfirm();
+          },
+          style: ElevatedButton.styleFrom(
+            backgroundColor: theme.colorScheme.error,
+            foregroundColor: theme.colorScheme.onError,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
+            ),
+          ),
+          child: Text(confirmButtonText),
+        ),
+      ],
+    );
+  }
+}
