+++ lib/presentation/widgets/common/empty_state_view.dart
@@ -1,0 +1,80 @@
+import 'package:flutter/material.dart';
+
+class EmptyStateView extends StatelessWidget {
+  final String title;
+  final String message;
+  final IconData icon;
+  final Color? iconColor;
+  final String? buttonText;
+  final VoidCallback? onButtonPressed;
+
+  const EmptyStateView({
+    required this.title,
+    required this.message,
+    required this.icon,
+    this.iconColor,
+    this.buttonText,
+    this.onButtonPressed,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final colorScheme = theme.colorScheme;
+    
+    return Center(
+      child: Padding(
+        padding: const EdgeInsets.all(32),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Container(
+              width: 80,
+              height: 80,
+              decoration: BoxDecoration(
+                color: (iconColor ?? colorScheme.primary).withOpacity(0.1),
+                shape: BoxShape.circle,
+              ),
+              child: Icon(
+                icon,
+                size: 40,
+                color: iconColor ?? colorScheme.primary,
+              ),
+            ),
+            const SizedBox(height: 24),
+            Text(
+              title,
+              style: theme.textTheme.titleLarge?.copyWith(
+                fontWeight: FontWeight.bold,
+              ),
+              textAlign: TextAlign.center,
+            ),
+            const SizedBox(height: 12),
+            Text(
+              message,
+              style: theme.textTheme.bodyMedium?.copyWith(
+                color: colorScheme.onSurface.withOpacity(0.7),
+              ),
+              textAlign: TextAlign.center,
+            ),
+            if (buttonText != null && onButtonPressed != null) ...[
+              const SizedBox(height: 32),
+              ElevatedButton.icon(
+                onPressed: onButtonPressed,
+                icon: const Icon(Icons.add),
+                label: Text(buttonText!),
+                style: ElevatedButton.styleFrom(
+                  padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                ),
+              ),
+            ],
+          ],
+        ),
+      ),
+    );
+  }
+}
