+++ lib/presentation/widgets/common/error_dialog.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+
+class ErrorDialog extends StatelessWidget {
+  final String title;
+  final String message;
+  final String buttonText;
+  final VoidCallback? onButtonPressed;
+
+  const ErrorDialog({
+    required this.title,
+    required this.message,
+    this.buttonText = 'OK',
+    this.onButtonPressed,
+    super.key,
+  });
+
+  static Future<void> show({
+    required BuildContext context,
+    required String title,
+    required String message,
+    String buttonText = 'OK',
+    VoidCallback? onButtonPressed,
+  }) async {
+    return showDialog<void>(
+      context: context,
+      builder: (context) => ErrorDialog(
+        title: title,
+        message: message,
+        buttonText: buttonText,
+        onButtonPressed: onButtonPressed,
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text(title),
+      content: Text(message),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(16),
+      ),
+      actions: [
+        TextButton(
+          onPressed: () {
+            Navigator.of(context).pop();
+            if (onButtonPressed != null) {
+              onButtonPressed!();
+            }
+          },
+          style: TextButton.styleFrom(
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(12),
+            ),
+          ),
+          child: Text(buttonText),
+        ),
+      ],
+    );
+  }
+}
