+++ lib/presentation/widgets/expenses/expense_card.dart
@@ -1,0 +1,173 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:mybudget/core/controllers/category_controller.dart';
+import 'package:mybudget/data/models/category_model.dart';
+import 'package:mybudget/data/models/expense_model.dart';
+import 'package:mybudget/presentation/widgets/common/delete_confirmation_dialog.dart';
+
+class ExpenseCard extends StatelessWidget {
+  final ExpenseModel expense;
+  final String accountName;
+  final VoidCallback onDelete;
+  final VoidCallback onEdit;
+
+  const ExpenseCard({
+    required this.expense,
+    required this.accountName,
+    required this.onDelete,
+    required this.onEdit,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Row(
+              children: [
+                Container(
+                  width: 40,
+                  height: 40,
+                  decoration: BoxDecoration(
+                    color: Theme.of(context).colorScheme.errorContainer,
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                  child: Icon(
+                    _getCategoryIcon(expense.categoryId),
+                    color: Theme.of(context).colorScheme.error,
+                  ),
+                ),
+                const SizedBox(width: 16),
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(
+                        expense.name,
+                        style: Theme.of(context).textTheme.titleMedium,
+                      ),
+                      const SizedBox(height: 4),
+                      Text(
+                        accountName,
+                        style: Theme.of(context).textTheme.bodySmall,
+                      ),
+                    ],
+                  ),
+                ),
+                Text(
+                  '${expense.amount.toStringAsFixed(2)} €',
+                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                    color: Theme.of(context).colorScheme.error,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(height: 12),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                if (expense.frequency != 'Unique')
+                  Chip(
+                    label: Text(expense.frequency),
+                    backgroundColor:
+                        Theme.of(context).colorScheme.primaryContainer,
+                    labelStyle: TextStyle(
+                      color: Theme.of(context).colorScheme.onPrimaryContainer,
+                    ),
+                  ),
+                const SizedBox(width: 8),
+                Chip(
+                  label: Text(_formatDate(expense.date, expense.frequency)),
+                  backgroundColor:
+                      Theme.of(context).colorScheme.secondaryContainer,
+                  labelStyle: TextStyle(
+                    color: Theme.of(context).colorScheme.onSecondaryContainer,
+                  ),
+                ),
+              ],
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [
+                IconButton(
+                  icon: const Icon(Icons.delete),
+                  onPressed: () {
+                    DeleteConfirmationDialog.show(
+                      context: context,
+                      title: 'Confirmer la suppression',
+                      message:
+                          'Voulez-vous vraiment supprimer ${expense.name} ?',
+                      onConfirm: onDelete,
+                    );
+                  },
+                ),
+                IconButton(icon: const Icon(Icons.edit), onPressed: onEdit),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  IconData _getCategoryIcon(int categoryId) {
+    final categoryController = Get.find<CategoryController>();
+    final category = categoryController.categories.firstWhere(
+      (c) => c.id == categoryId,
+      orElse: () => CategoryModel()..icon = 'category',
+    );
+
+    switch (category.icon) {
+      case 'restaurant':
+        return Icons.restaurant;
+      case 'directions_car':
+        return Icons.directions_car;
+      case 'home':
+        return Icons.home;
+      case 'sports_esports':
+        return Icons.sports_esports;
+      case 'medical_services':
+        return Icons.medical_services;
+      case 'shopping_bag':
+        return Icons.shopping_bag;
+      default:
+        return Icons.category;
+    }
+  }
+
+  String _formatDate(DateTime date, String frequency) {
+    switch (frequency) {
+      case 'Unique':
+        return '${date.day}/${date.month}/${date.year}';
+      case 'Mensuel':
+        return 'Jour ${date.day}';
+      case 'Hebdomadaire':
+        final days = ['Dim', 'Lun', 'Mar', 'Mer', 'Jeu', 'Ven', 'Sam'];
+        return days[date.weekday % 7];
+      case 'Annuel':
+        final months = [
+          'Jan',
+          'Fév',
+          'Mar',
+          'Avr',
+          'Mai',
+          'Juin',
+          'Juil',
+          'Août',
+          'Sep',
+          'Oct',
+          'Nov',
+          'Déc',
+        ];
+        return '${date.day} ${months[date.month - 1]}';
+      default:
+        return '${date.day}/${date.month}/${date.year}';
+    }
+  }
+}
